Feat: 카드 디자인 수정, 페이지 연결
