+++ src/pages/placeDetail/components/MapModal.tsx
@@ -45,8 +45,9 @@ const customStyles = {
     left: '50%',
     right: 'auto',
     bottom: 'auto',
-    marginRight: '-50%',
     transform: 'translate(-50%, -50%)',
+    borderRadius: '10px',
+    padding: '2.5rem',
   },
 };
