--- src/pages/reservation/components/ReservationAgree.tsx
@@ -1,73 +1,0 @@
-import { Checkbox } from '@mui/material';
-import { theme } from '@styles/theme';
-import styled from 'styled-components';
-
-const ReservationAgree = () => {
-  return (
-    <StyledRuleWrapper>
-      <StyledEssentialTerms>
-        <span>필수 약관 동의</span>
-        <StyledEssentialCheckList>
-          <CheckboxContainer>
-            <Checkbox />
-            <CheckboxText>[필수]개인정보 제 3자 제공</CheckboxText>
-          </CheckboxContainer>
-          <CheckboxContainer>
-            <Checkbox />
-            <CheckboxText>[필수]만 14세 이상 이용 동의</CheckboxText>
-          </CheckboxContainer>
-          <CheckboxContainer>
-            <Checkbox />
-            <CheckboxText>[필수]개인정보 수집 및 이용</CheckboxText>
-          </CheckboxContainer>
-          <StyledPayText>
-            <span>※ 이용규칙, 취소 및 환불 규칙</span>에 동의하실 경우
-            결제하기를 클릭해주세요
-          </StyledPayText>
-        </StyledEssentialCheckList>
-      </StyledEssentialTerms>
-    </StyledRuleWrapper>
-  );
-};
-export default ReservationAgree;
-
-export const StyledRuleWrapper = styled.div`
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  margin-left: 0.4rem;
-  margin-top: 2rem;
-  /* border: 1px solid black; */
-`;
-
-export const StyledEssentialTerms = styled.div`
-  /* border: 1px solid black; */
-  margin: 2rem;
-  text-align: start;
-  span {
-    padding-left: 0.75rem;
-    font-weight: bold;
-    color: ${theme.colors.navy};
-  }
-`;
-
-export const StyledEssentialCheckList = styled.div``;
-
-export const StyledPayText = styled.p`
-  text-align: start;
-  margin-left: 0.75rem;
-  span {
-    color: red;
-    font-weight: bold;
-  }
-`;
-
-export const CheckboxContainer = styled.div`
-  display: flex;
-`;
-
-export const CheckboxText = styled.p`
-  line-height: 2.65rem;
-  font-weight: bold;
-  font-size: 0.95rem;
-`;
